[Views] merge with main updates
