+++ src/semantic_world/world_rdr/world_views_mcrdr_cases.py
@@ -1,0 +1,74 @@
+# This file contains the corner cases for the rules.
+try:
+    from test.test_views.test_views import ViewTestCase
+except  ImportError:
+    ViewTestCase = None
+from ripple_down_rules.datastructures.dataclasses import CaseFactoryMetaData
+
+
+corner_case_90574698325129464513441443063592862114 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_apartment_views')
+
+
+corner_case_14920098271685635920637692283091167284 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_apartment_views')
+
+
+corner_case_331345798360792447350644865254855982739 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_apartment_views')
+
+
+corner_case_35528769484583703815352905256802298589 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_apartment_views')
+
+
+corner_case_59112619694893607910753808758642808601 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_door_view')
+
+
+corner_case_10840634078579061471470540436169882059 = CaseFactoryMetaData(factory_method='ViewTestCase.get_kitchen_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_fridge_view')
+
+
+corner_case_298893002154434174914882697804321727044 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_roots_view')
+
+
+corner_case_95749960275749164559369305418046530921 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_table_view')
+
+
+corner_case_56044667384119005486660532922350093576 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_windows_view')
+
+
+corner_case_192414996332569377000417356221731692718 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_walls_view')
+
+
+corner_case_295145914874381034058732290061016216922 = CaseFactoryMetaData(factory_method='ViewTestCase.get_table_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_surface_view')
+
+
+corner_case_76734351718414165115161689080544922137 = CaseFactoryMetaData(factory_method='ViewTestCase.get_table_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_leg_view')
+
+
+corner_case_8779700588076492791914944718923983853 = CaseFactoryMetaData(factory_method='ViewTestCase.get_table_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_DetailedTable_view')
+
+
+corner_case_276870538550900219151385672580612993791 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_Armchair_view')
+
+
+corner_case_272346104137726532028618426380451371764 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_sofa_view')
+
+
+corner_case_125114902106458283648997265474531641579 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_oven_view')
+
+
+corner_case_39216173054479960152579027281679377283 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_hotplate_view')
+
+
+corner_case_56932022228577464889761478858979829689 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_sides_view')
+
+
+corner_case_304562387165012343799590402190050617477 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_countertop_view')
+
+
+corner_case_49636263552498546597487129028109645395 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_cooktop_view')
+
+
+corner_case_335129869970075176340711076346427700506 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_DetailedCooktop_view')
+
+
+corner_case_180013694937072261831339544475640316362 = CaseFactoryMetaData(factory_method='ViewTestCase.get_apartment_world', factory_idx=None, case_conf=None, scenario='ViewTestCase.test_sink_view')
+
+
